DataScience_Project2/weather and energy resource/requirements.txt
pandas>=2.1
requests>=2.31
tenacity>=8.2
python-dateutil>=2.8
pyarrow>=12
